+++ apps/auth-server/src/EShopOnAbp.AuthServer/Program.cs
@@ -2,8 +2,6 @@ using System;
 using System.Threading.Tasks;
 using EShopOnAbp.Shared.Hosting.AspNetCore;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 using Serilog;
 
 namespace EShopOnAbp.AuthServer;
@@ -19,14 +17,7 @@ public class Program
         try
         {
             Log.Information($"Starting {assemblyName}.");
-            var builder = WebApplication.CreateBuilder(args);
-            builder.Host
-                .AddAppSettingsSecretsJson()
-                .UseAutofac()
-                .UseSerilog();
-
-            await builder.AddApplicationAsync<EShopOnAbpAuthServerModule>();
-            var app = builder.Build();
+            var app = await ApplicationBuilderHelper.BuildApplicationAsync<EShopOnAbpAuthServerModule>(args);
             await app.InitializeApplicationAsync();
             await app.RunAsync();
