+++ apps/public-web/src/EShopOnAbp.PublicWeb/Program.cs
@@ -2,8 +2,6 @@ using System;
 using System.Threading.Tasks;
 using EShopOnAbp.Shared.Hosting.AspNetCore;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 using Serilog;
 
 namespace EShopOnAbp.PublicWeb;
@@ -19,17 +17,10 @@ public class Program
         try
         {
             Log.Information($"Starting {assemblyName}.");
-            var builder = WebApplication.CreateBuilder(args);
-            builder.Host
-                .AddAppSettingsSecretsJson()
-                .UseAutofac()
-                .UseSerilog();
-
-            await builder.AddApplicationAsync<EShopOnAbpPublicWebModule>();
-            var app = builder.Build();
+            var app = await ApplicationBuilderHelper.BuildApplicationAsync<EShopOnAbpPublicWebModule>(args);
             await app.InitializeApplicationAsync();
             await app.RunAsync();
-            
+
             return 0;
         }
         catch (Exception ex)
